refactor: copy awsconfig into amplifyconfig
Create an auth block that contains the contents of awsconfig
